Replace setExact and setAlarmClock with setAndAllowWhileIdle
On Android 13 and up, exact alarms are not allowed by default:
https://developer.android.com/about/versions/14/changes/schedule-exact-alarms
One has to ask the user to change a setting to allow them. This works
less well for us than making the alarms inexact.

Replacing the upstream code is crude. Ideally I would add a parameter
for allowing alarms to be inexact. But that would require dealing with
the plugin code to a far greater extent than I'm willing to. I would
rather write a new bare-bones plugin that does the minimum amount of
work needed to suit our purpose. But throwing out a few lines of
upstream code and changing the method it calls in two places is even
easier.